rename TiffHeader to tiff.Header
