+++ exif/reader_test.go
@@ -31,13 +31,13 @@ func TestExifReader(t *testing.T) {
 	}
 
 	// SetHeader
-	th := tiff.NewTiffHeader(byteOrder, exifOffset, exifOffset, 0)
+	th := tiff.NewHeader(byteOrder, exifOffset, exifOffset, 0)
 	if err := er.SetHeader(th); err != nil {
 		t.Errorf("Error with reader.SetHeader expected no error")
 	}
 
 	// SetHeader (Invalid Header)
-	if err := er.SetHeader(tiff.TiffHeader{}); err != ErrInvalidHeader {
+	if err := er.SetHeader(tiff.Header{}); err != ErrInvalidHeader {
 		t.Errorf("Error with reader.SetHeader %s, expected %s", err.Error(), ErrInvalidHeader.Error())
 	}
