+++ tiff/tiff_meta.go
@@ -1,15 +1,15 @@
 package tiff
 
 type TiffMetadata struct {
-	TiffHeader
+	header Header
 }
 
 func (tm TiffMetadata) Size() (width uint16, height uint16) {
 	return 0, 0
 }
 
-func (tm TiffMetadata) Header() TiffHeader {
-	return tm.TiffHeader
+func (tm TiffMetadata) Header() Header {
+	return tm.header
 }
 
 func (tm TiffMetadata) XMP() string {
