--- xmpheader.go
@@ -1,7 +1,0 @@
-package imagemeta
-
-// XMPHeader -
-type XMPHeader struct {
-	XMPoffset uint32
-	XMPlength uint32
-}
